[core] split out test that needs java (#57007)
and add `needs_java` tag to the test. this will allow us to run "normal"
tests without java jdk and jre, where and running java related tests
seperately
